Adjusted Colliders
Final bug fix for Dummy's Dilemma. Fixed the collisions so that the player character is always on the ground/platform and not floating above platforms or corpse platforms. Also adjusted the yOffset of the corpse platform spawning to prevent the player from dying twice if the player's negative y velocity was too fast and the hitbox clipped through the corpse platform